refactor(gateway): extract silence-poke wiring (#2996 P8 PR-C3)
Move the silencePoke.startTimer options object VERBATIM from gateway.ts into
telegram-plugin/gateway/liveness-wiring.ts (buildSilencePokeOptions): the
floorState role reader, the approval-blocked onMidTurnFloor re-ping, and the
300 s onFrameworkFallback unwedge (turn teardown, stale sibling purge via
purgeStaleTurnsForChat, PR-4e keyed clear + purgeChatStale self-heal, buffer
self-heal redeliver) — behind a LivenessWiringDeps builder.

- P0c: the module exports the options body; the GATEWAY still registers the
  timer (isGatewayMain-gated boot side effect unchanged).
- The two user-visible sends cross as the injected sendSilenceText closure —
  the bot.api touch stays behind the gateway retry-policy allowlist; the
  loud/silent gate (blockedOnApproval) rides the closure's silent argument.
- Volatile reads (currentTurn, inFlightUpdate, turnsDb, inboundSpool,
  ipcServer) cross as getters (P9 getter funnel, clusters G/H).
- Source-scrape suites retargeted in the same PR: per-topic-current-turn
  (keyed delete + purgeChatStale oracles), no-reply-bounded-drain (fallback
  long-stop literal), silence-liveness-wiring (notification gate).
- gateway-line-ratchet lowered 25900 -> 25580 (gateway.ts now 25532 lines).

Refs #2996. Job: reference/jobs/on-the-leash.md.
