+++ scripts/gateway-line-ratchet.txt
@@ -1,1 +1,1 @@
-25900
+25580
